add mailing

+++ endpoint.py
@@ -1,4 +1,4 @@
-from fastapi import Depends, HTTPException, status, APIRouter
+from fastapi import Depends, HTTPException, status
 from fastapi_pagination import paginate
 from fastapi import FastAPI
 from fastapi_pagination.bases import AbstractPage
@@ -25,7 +25,6 @@ async def user_add(data: UserSchema, db: Session = Depends(get_db)) -> UserSchem
 
 @app.post('/subscribe', status_code=201, response_model=SubscribeResponseSchema, tags=['subscribe'])
 async def user_subscribe(data: SubscribeSchema, db: Session = Depends(get_db)) -> SubscribeResponseSchema:
-
     try:
         user = db.query(User).filter(User.id == data.user_id).one()
     except NoResultFound:
@@ -38,6 +37,7 @@ async def user_subscribe(data: SubscribeSchema, db: Session = Depends(get_db)) -
     try:
         db.add(user)
         db.commit()
+
     except IntegrityError:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
@@ -47,23 +47,30 @@ async def user_subscribe(data: SubscribeSchema, db: Session = Depends(get_db)) -
     return subscribe
 
 
-@app.delete('/unsubscribe', status_code=200, response_model=SubscribeResponseSchema, tags=['subscribe'])
+@app.delete('/unsubscribe', status_code=200, response_model=list[SubscribeResponseSchema], tags=['subscribe'])
 async def user_subscribe(data: SubscribeSchema, db: Session = Depends(get_db)) -> SubscribeResponseSchema:
-    subscribe = db.query(Subscribe).filter(Subscribe.user_id == data.user_id,
-                                           Subscribe.blog_id == data.blog_id).one()
-
-    db.delete(subscribe)
+    subscribes = db.query(Subscribe).filter(Subscribe.user_id == data.user_id,
+                                            Subscribe.blog_id == data.blog_id).all()
+    if len(subscribes) == 0:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail='Subscribe not found'
+        )
+    for subscribe in subscribes:
+        db.delete(subscribe)
     db.commit()
 
-    return subscribe
+    return subscribes
 
 
 @app.get('/tape/{user_id}', status_code=200, response_model=Page[PostSchema], tags=['tape'])
 async def tape(user_id: int, post_id: int | None = None, db: Session = Depends(get_db)) -> AbstractPage[PostSchema]:
-    if post_id is not None:
-        await redis.lpush(user_id, post_id)
     try:
         user = db.query(User).filter(User.id == user_id).one()
+
+        if post_id is not None:
+            await redis.lpush(user_id, post_id)
+
     except NoResultFound:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
@@ -81,12 +88,34 @@ async def tape(user_id: int, post_id: int | None = None, db: Session = Depends(g
     return paginate(db.query(Post).filter(Post.id.in_(posts)).order_by(Post.created_at).all())
 
 
-@app.get('/all/{user_id}', status_code=200)
-async def all_subscribe(user_id: int, db: Session = Depends(get_db)):
-    user = db.query(User).filter(User.id == user_id).one()
-    res = []
-    for i in user.subscribers:
-        res.append({'user_id': i.user_id,
-                    'blog_id': i.blog_id})
+@app.get('/tape/limit_5/{user_id}', status_code=200, response_model=Page[PostSchema], tags=['tape'])
+async def tape(user_id: int, post_id: int | None = None, db: Session = Depends(get_db)) -> AbstractPage[PostSchema]:
+    try:
+        user = db.query(User).filter(User.id == user_id).one()
 
-    return res
+        if post_id is not None:
+            await redis.lpush(user_id, post_id)
+
+    except NoResultFound:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail='User not found'
+        )
+    posts = []
+    cache = await redis.lrange(user_id, 0, -1)
+
+    for subscribe in user.subscribers:
+        for post in subscribe.blog.post:
+            if str(post.id) in cache:
+                continue
+            posts.append(post.id)
+
+    return paginate(db.query(Post).filter(Post.id.in_(posts)).order_by(Post.created_at).limit(5).all())
+
+
+@app.get('/user/all', status_code=200, response_model=list[UserResponseSchema], tags=['subscribe'])
+async def subscribe_all(db: Session = Depends(get_db)):
+    users = db.query(User).order_by(User.id).all()
+    return users
+
+
